Replaced React.error/React.warn with React.getComponentStack
